python-service/requirements.txt
fastapi
uvicorn
earthengine-api
numpy
scikit-learn
pydantic
google-auth
pandas
